Removes the video from the disc
Removes the removed video also from the disc. Verifies that the uploaded video
has finished the conversion for its removal.

+++ service/src/main/java/org/sing_group/piba/service/video/DefaultVideoService.java
@@ -112,7 +112,11 @@ public class DefaultVideoService implements VideoService {
 
   @Override
   public void delete(Video video) {
+    if (video.isProcessing()) {
+      throw new IllegalArgumentException("Can not be deleted until the conversion is completed");
+    }
     this.videoDao.delete(video);
+    this.videoStorage.delete(video.getId());
   }
 
   @Override
